comment+vote: moved code into its own PHP file
